backend/requirements.txt
# Essential packages for the resume agent backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6

# PDF generation libraries
reportlab==4.0.7
weasyprint==60.2
pypdf2==3.0.1

# AI/ML libraries
openai==1.3.7
anthropic==0.7.8
transformers==4.36.0
torch==2.1.1

# Database
sqlalchemy==2.0.23

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
